+++ modules/apps/commerce/commerce-inventory-api/src/main/java/com/liferay/commerce/inventory/service/CommerceInventoryReplenishmentItemService.java
@@ -15,6 +15,7 @@ import com.liferay.portal.kernel.transaction.Isolation;
 import com.liferay.portal.kernel.transaction.Propagation;
 import com.liferay.portal.kernel.transaction.Transactional;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -45,9 +46,9 @@ public interface CommerceInventoryReplenishmentItemService extends BaseService {
 	 */
 	public CommerceInventoryReplenishmentItem
 			addCommerceInventoryReplenishmentItem(
-				String externalReferenceCode, long commerceInventoryWarehouseId,
-				Date availabilityDate, int quantity, String sku,
-				String unitOfMeasureKey)
+			String externalReferenceCode, long commerceInventoryWarehouseId,
+			Date availabilityDate, BigDecimal quantity, String sku,
+			String unitOfMeasureKey)
 		throws PortalException;
 
 	public void deleteCommerceInventoryReplenishmentItem(
@@ -107,9 +108,9 @@ public interface CommerceInventoryReplenishmentItemService extends BaseService {
 
 	public CommerceInventoryReplenishmentItem
 			updateCommerceInventoryReplenishmentItem(
-				String externalReferenceCode,
-				long commerceInventoryReplenishmentItemId,
-				Date availabilityDate, int quantity, long mvccVersion)
+			String externalReferenceCode,
+			long commerceInventoryReplenishmentItemId,
+			Date availabilityDate, BigDecimal quantity, long mvccVersion)
 		throws PortalException;
 
 }
